requirements.txt
streamlit==1.28.0
pandas==2.1.3
pyspellchecker==0.7.2
language-tool-python==2.7.1
openpyxl==3.1.2
